Harden native ethash large-file seeking

+++ src/Native/libetchash/io.h
@@ -79,7 +79,7 @@ enum ethash_io_rc {
  * @param origin       Initial position
  * @return             Current offset or -1 to indicate an error
  */
-int ethash_fseek(FILE* f, size_t offset, int origin);
+int ethash_fseek(FILE* f, uint64_t offset, int origin);
 
 /**
  * Prepares io for ethash
